Added streaks and counters
Streaks are when correct answers are simultaneous. When more than 0
answers are answered correctly simultaneously, the last element of the
"streaks" array which is a in the Round class / objects is incremented
by 1, provided that the streaks array stores numbers. When the question
is answered incorrectly or skipped the streak is cut and a new element
is pushed to the "streaks" array. Additionally, counters are what
displays for the users how many correct answers were submitted and for
the current streak.