Made the project modular
As stated, the project was made modular instead of having a single
1000-line file holding everything. the directories added are: 1- The
scripts directory, which contains three directories (dirs): the app dir,
which contains the files that make up the app, the utils dir, which
contains functions that are reused, and the shared dir, which contains
data shared used among other dirs, it, namely, contains the words.js
file. 2- The styles dir, which contains the style.css file and a scss
dir which consists of various scss files and, furthermore, consists of
the partials dir, which contains partial scss files (variables, mixins).
3- The icons dir, which, as the name suggests, contains all of the icons
used in the project. Additionally, the project got npm initialized.A
fair amount of minor changes, moreover, were made.