Added forms and options modification
Forms are added for options and the answer textbox. Now when the restart
is styled when having the data attribute end-round. Most crucially,
the modal was styled and made more appealing to look at. It was also
made responsive. Furthermore, the options' menu was modified, making the
portion container scroll, instead of the modal, and modifying other
modal's element's styles. That's alongside adding a form element, which
is actually what scrolls. Additionally, a close button was added.