Database insertion
the "top" and "habitat" so-called branches were added. stopwatch
functions, moreover, were slightly modified to not reduce the effect
on performance and not have weird behaviors (when stopwatch is off, it
won't change DOM, unlike what it did before). Additionally. fixed a bug
that caused rounds to run show the "congrats" section at the beginning
of a round sometimes.